Add files via upload
This machine learning project generates an approximate calculation of the cost of Apple iPhone based on the trajectory of the price of different models over the years. We use the .csv file to generate an analysis of iPhone prices and how they increase through LinearRegression( ideal model ) of the increases of Apple iPhones without included external factors in the cost of models and apples manufacturing cost. This is purely based on the cost of iPhones in the last couple years. The .py file includes the code of the project, build using python and three different python packages : scikit-learn, pandas and matplotlib.